+++ src/test/java/com/keeper/homepage/global/dsl/rest_docs/RestDocsRequestBuilder.kt
@@ -8,19 +8,12 @@ import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilde
 import org.springframework.util.Assert
 import java.net.URI
 
-class RestDocsRequestBuilder {
-
-    private val mockMvc: MockMvc
-    private val method: DocsMethod
-    private val url: String
-    private val pathParams: Array<*>?
-
-    constructor(mockMvc: MockMvc, method: DocsMethod, url: String, pathParams: Array<*>? = null) {
-        this.method = method
-        this.url = url
-        this.mockMvc = mockMvc
-        this.pathParams = pathParams
-    }
+class RestDocsRequestBuilder(
+        private val mockMvc: MockMvc,
+        private val method: DocsMethod,
+        private val url: String,
+        private val pathParams: Array<*>? = null
+) {
 
     fun build(): MockHttpServletRequestBuilder? {
         val mockRequestBuilder = when (method) {
